+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+<a name="0.19.9"></a>
+## [0.19.9](https://github.com/angular-ui/ui-select/compare/v0.19.8...v0.19.9) (2019-04-25)
+
+### Bug Fixes
+
+* **uiSelectDirective:** Timeout reposition calculation due to child directives not updated yet in angular.js 1.5.
+  [VIV-730](https://github.com/bdb-opensource/ui-select/pull/8)
+* **uiSelectChoices:** Remove ng-if="$select.isOpen" from rows template since it has bad performance in large lists.
+  [VIV-730](https://github.com/bdb-opensource/ui-select/pull/8)
+* Fix positioning for left or right positioning of the dropdown menu.
+  [AB-7708](https://github.com/bdb-opensource/ui-select/pull/6)
+* Remove position of detached select [AB-7754](https://github.com/bdb-opensource/ui-select/pull/5)
+* **uiSelectController:** Use capturing event to close the dropdown
+  [AB-7284](https://github.com/bdb-opensource/ui-select/pull/4)
+
+### Features
+
+* **uiSelectChoices:** Remove clear button for null value entry, uses null-label and null-value properties for single
+  select templates. [VIV-730](https://github.com/bdb-opensource/ui-select/pull/5)
+
 <a name="0.19.8"></a>
 ## [0.19.8](https://github.com/angular-ui/ui-select/compare/v0.19.7...v0.19.8) (2017-04-15)